Changed Texture Ordering, Byte to Int and Texture Order
- Reverted the change to existing texture placements in the array to not break already existing multis;
- Changed the output value type of getTextureCasingIndex to int, to support the intended texture.

+++ src/main/java/gregtech/common/tileentities/machines/multi/GT_MetaTileEntity_LargeTurbine_HPSteam.java
@@ -79,7 +79,7 @@ public class GT_MetaTileEntity_LargeTurbine_HPSteam extends GT_MetaTileEntity_La
     }
 
     @Override
-    public byte getCasingTextureIndex() {
+    public int getCasingTextureIndex() {
         return 59;
     }
